+++ src/Component/Collector.php
@@ -18,8 +18,8 @@ class Collector
      * Collects the specified components.
      *
      * @param  \Interop\Container\ContainerInterface $container
-     * @param  array|null                            $globals
      * @param  array                                 $components
+     * @param  array|null                            $globals
      * @return \Rougin\Slytherin\Component\Collection
      */
     public static function get(ContainerInterface $container, $components = [], &$globals = null)
@@ -55,6 +55,7 @@ class Collector
 
         $collection->setContainer($container);
 
+        // NOTE: To be removed in v1.0.0
         if ($globals) {
             $globals['container'] = $container;
         }