+++ tests/Fixture/Components/DebuggerComponent.php
@@ -23,6 +23,7 @@ class DebuggerComponent extends AbstractComponent
 
     /**
      * Returns an instance from the named class.
+     * It's used in supporting component types for Slytherin.
      *
      * @return mixed
      */