+++ tests/Fixture/Components/DispatcherComponent.php
@@ -24,6 +24,7 @@ class DispatcherComponent extends AbstractComponent
 
     /**
      * Returns an instance from the named class.
+     * It's used in supporting component types for Slytherin.
      *
      * @return mixed
      */