+++ tests/Fixture/Components/MiddlewareComponent.php
@@ -25,6 +25,7 @@ class MiddlewareComponent extends AbstractComponent
 
     /**
      * Returns an instance from the named class.
+     * It's used in supporting component types for Slytherin.
      *
      * @return mixed
      */