+++ tests/Fixture/Components/SingleComponent.php
@@ -13,20 +13,4 @@ use Rougin\Slytherin\Component\AbstractComponent;
  */
 class SingleComponent extends AbstractComponent
 {
-    /**
-     * Name of the class to be added in the container.
-     *
-     * @var string
-     */
-    protected $className = 'Rougin\Slytherin\Test\Fixture\TestClass';
-
-    /**
-     * Returns an instance from the named class.
-     *
-     * @return mixed
-     */
-    public function get()
-    {
-        return new TestClass;
-    }
 }